requirements.txt

fastapi 
pydantic

google-adk 
a2a-sdk 
uvicorn 
httpx 
python-dotenv 
rich
autogen-agentchat 
autogen-ext 
langchain-core 
langgraph
oci 
litellm